// src/lib/edgeql-interfaces.ts
export namespace std {
  export interface BaseObject {
    "id": string;
  }
  export interface $Object extends BaseObject {}
  export type Endian = "Little" | "Big";
  export interface FreeObject extends BaseObject {}
  export type JsonEmpty = "ReturnEmpty" | "ReturnTarget" | "Error" | "UseNull" | "DeleteKey";
  export namespace enc {
    export type Base64Alphabet = "standard" | "urlsafe";
  }
}
// ...
  export interface Course extends std.$Object {
    "code": string;
    "description": string;
    "label": string;
  }
  export interface Osteopath extends std.$Object {
    "username": string;
    "student"?: Student | null;
    "bio"?: string | null;
    "friday"?: unknown | null;
    "monday"?: unknown | null;
    "saturday"?: unknown | null;
    "sunday"?: unknown | null;
    "thursday"?: unknown | null;
    "tuesday"?: unknown | null;
    "verification_by"?: string | null;
    "wednesday"?: unknown | null;
    "user"?: User | null;
  }
  export type Role = "admin" | "user" | "osteopath";
  export interface Session extends std.$Object {
    "user": User;
    "expiresAt": Date;
  }
  export interface Student extends std.$Object {
    "course": Course;
    "email": string;
    "batch"?: string | null;
    "isAlumni"?: boolean | null;
    "university"?: string | null;
  }
  export interface User extends std.$Object {
    "osteopath"?: Osteopath | null;
    "email": string;
    "createdAt"?: Date | null;
    "name"?: string | null;
    "phoneNumber"?: string | null;
    "picture"?: string | null;
    "role"?: Role | null;
    "status"?: string | null;
    "updatedAt"?: Date | null;
    "sessions": Session[];
  }